feat(irs-api):[TRI-497] Changed package name from assemblypartrelationship to assetadministrationshell
