+++ irs-api/src/main/java/net/catenax/irs/aaswrapper/job/ItemContainer.java
@@ -16,7 +16,7 @@ import lombok.Getter;
 import lombok.Singular;
 import lombok.extern.jackson.Jacksonized;
 import net.catenax.irs.component.Tombstone;
-import net.catenax.irs.component.assemblypartrelationship.AssetAdministrationShellDescriptor;
+import net.catenax.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import net.catenax.irs.dto.AssemblyPartRelationshipDTO;
 
 /**
