+++ irs-api/src/main/java/net/catenax/irs/aaswrapper/registry/domain/DigitalTwinRegistryClient.java
@@ -13,7 +13,7 @@ import static net.catenax.irs.configuration.RestTemplateConfig.OAUTH_REST_TEMPLA
 
 import java.net.URI;
 
-import net.catenax.irs.component.assemblypartrelationship.AssetAdministrationShellDescriptor;
+import net.catenax.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
