+++ irs-api/src/main/java/net/catenax/irs/aaswrapper/registry/domain/DigitalTwinRegistryFacade.java
@@ -15,8 +15,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.catenax.irs.aaswrapper.job.AspectTypeFilter;
-import net.catenax.irs.component.assemblypartrelationship.AssetAdministrationShellDescriptor;
-import net.catenax.irs.component.assemblypartrelationship.SubmodelDescriptor;
+import net.catenax.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
+import net.catenax.irs.component.assetadministrationshell.SubmodelDescriptor;
 import net.catenax.irs.component.enums.AspectType;
 import net.catenax.irs.dto.JobParameter;
 import org.springframework.stereotype.Service;
