+++ irs-api/src/main/java/net/catenax/irs/configuration/OpenApiExamples.java
@@ -31,13 +31,13 @@ import net.catenax.irs.component.QueryParameter;
 import net.catenax.irs.component.Relationship;
 import net.catenax.irs.component.Summary;
 import net.catenax.irs.component.Tombstone;
-import net.catenax.irs.component.assemblypartrelationship.AssetAdministrationShellDescriptor;
-import net.catenax.irs.component.assemblypartrelationship.Endpoint;
-import net.catenax.irs.component.assemblypartrelationship.IdentifierKeyValuePair;
-import net.catenax.irs.component.assemblypartrelationship.LangString;
-import net.catenax.irs.component.assemblypartrelationship.ProtocolInformation;
-import net.catenax.irs.component.assemblypartrelationship.Reference;
-import net.catenax.irs.component.assemblypartrelationship.SubmodelDescriptor;
+import net.catenax.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
+import net.catenax.irs.component.assetadministrationshell.Endpoint;
+import net.catenax.irs.component.assetadministrationshell.IdentifierKeyValuePair;
+import net.catenax.irs.component.assetadministrationshell.LangString;
+import net.catenax.irs.component.assetadministrationshell.ProtocolInformation;
+import net.catenax.irs.component.assetadministrationshell.Reference;
+import net.catenax.irs.component.assetadministrationshell.SubmodelDescriptor;
 import net.catenax.irs.component.enums.AspectType;
 import net.catenax.irs.component.enums.BomLifecycle;
 import net.catenax.irs.component.enums.Direction;
