+++ irs-api/src/test/java/net/catenax/irs/aaswrapper/registry/domain/DigitalTwinRegistryClientImplTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import net.catenax.irs.component.assemblypartrelationship.AssetAdministrationShellDescriptor;
+import net.catenax.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.client.RestTemplate;
